ai_module/requirements.txt
fastapi
uvicorn[standard]
transformers
requests
Pillow
pydantic
torch